requirements.txt
yfinance>=0.2.40
pandas>=2.0
numpy>=1.24
scipy>=1.11
python-dotenv>=1.0
requests>=2.31
pytz>=2023.3
